Disable Terminal while closed to reduce GC allocations
Unity has a quirk that even an empty OnGUI method allocates memory every
frame (see link below). This had an effect in our game such that the
console was triggering GC.Collect every 2-3 seconds, while it's closed.

This change introduces another component that'll listen for console
toggle keypresses in Update instead of OnGUI, and have the Terminal
disable itself when it's closed (so that OnGUI doesn’t get called).

https://forum.unity.com/threads/gc-collect-in-guiutility-begingui.642229/#post-4306870