Replaced autocomplete logic, pressing Tab multiple times cycles through
In the previous few commits I broke the usability because if there were multiple
completions available, nothing would get added to the input field.

This brings that back, and also improves it by allowing you to cycle through
completions by pressing Tab multiple times, and also replaces how the completion
results are displayed - they don't spam the log anymore, a different buffer is
drawn instead temporarily, while you're in "cycling mode".

Autocompletions are sorted so that commands that start with the input (or have
the input closer to the beginning) get shown first.