Initialize in Awake instead of Start, execution order -999
This aims to fix early Debug.Logs not appearing in console because they were
logged before CommandTerminal got initialized.

I chose the execution order to be -999 because I saw in a new project that
UnityEngine.EventSystems.EventSystem uses -1000 so I made it one after that just
in case, I didn't actually test anything related to EventSystem.